Refactored all commands out into their own files. Makes it easier to navigate that structure.
