+++ src/civilization/console/mod.rs
@@ -1,1 +1,12 @@
 pub mod console_commands;
+mod build_city_command;
+mod list_builds_command;
+mod show_board_command;
+mod end_move_command;
+mod move_command;
+mod list_moves_command;
+mod expand_population_command;
+mod start_command;
+mod player_end_building_command;
+mod eliminate_city_command;
+mod stupid_ai_command;
